requirements.txt
certifi==2018.4.16
chardet==3.0.4
cssselect==1.0.3
idna==2.7
lxml==4.2.3
numpy==1.14.5
oauthlib==2.1.0
pendulum==2.0.2
psycopg2==2.7.5
PySocks==1.6.8
python-dateutil==2.7.3
pytzdata==2018.5
readability-lxml==0.7
requests==2.19.1
requests-oauthlib==1.0.0
scipy==1.1.0
six==1.11.0
summa==1.0.0
tqdm==4.23.4
tweepy==3.5.0
urllib3==1.23
